requirements.txt
glog>=0.3.1 
mock>=2.0.0 
prometheus_client>=0.4.3
speedtest-cli>=2.0.2
